Switch to CFFI out-of-line ABI mode
