+++ MANIFEST.in
@@ -1,4 +1,5 @@
 include LICENSE
+include jack_build.py
 include *.rst
 include doc/requirements.txt
 recursive-include doc *.rst *.py
