+++ setup.py
@@ -29,8 +29,10 @@ setup(
     version=__version__,
     package_dir={'': 'src'},
     py_modules=['jack'],
-    install_requires=['CFFI'],
+    setup_requires=['CFFI>=1.0'],
+    install_requires=['CFFI>=1.0'],
     extras_require={'NumPy': ['NumPy']},
+    cffi_modules=['jack_build.py:ffibuilder'],
     author='Matthias Geier',
     author_email='Matthias.Geier@gmail.com',
     description='JACK Audio Connection Kit (JACK) Client for Python',
